refactor code for rest-api-1
